fix(types) strip TypeMeta from non-object types
The types used for plugin CRD ConfigFrom fields were originally written
as Kubernetes objects with TypeMeta, with validation rules set to ignore
the absence of TypeMeta fields.

As best I recall, I originally added TypeMeta to meet the
"// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object"
requirements. That deepcopy generator only works for Kubernetes objects,
so it fails if TypeMeta is absent. We have other deepcopy generators
available that generate the non-object deepcopy functions, which are all
we need.

Remove validation annotations from ConfigFrom types. With the removal of
TypeMeta, only the required fields are left, so we can remove the
optional annotation on the type and required annotation from individual
required fields.

Removes a case in the store for handling ConfigSource GVKs. These types
would never be deployed as standalone resources and this case was
unnecessary.